admin update all host request

+++ server/src/controller/seller.controller.js
@@ -1,7 +1,9 @@
-const { SellerCollection } = require("../module/module");
+const { SellerCollection, UserCollection } = require("../module/module");
 const catchAsync = require("../utils/catchAsync");
 const { ObjectId } = require("mongodb");
 const sendResponse = require("../utils/sendResponse");
+const { sendSellerStatusEmail } = require("../utils/sendMail");
+const AppError = require("../utils/AppError");
 
 const getAllSeller = catchAsync(async (req, res) => {
     const result = await SellerCollection.find().toArray();
@@ -44,6 +46,16 @@ const getSingleSellerByEmail = catchAsync(async (req, res) => {
 const deleteSeller = catchAsync(async (req, res) => {
     const id = req.params.id;
     const query = { _id: new ObjectId(id) };
+
+    const sellerRequest = await SellerCollection.findOne(query);
+    if (!sellerRequest) {
+        throw new AppError(404, "Seller request not found");
+    }
+
+    if (req.user.role !== "admin" && sellerRequest.sellerEmail !== req.user.email) {
+        throw new AppError(403, "You do not have permission to delete this request");
+    }
+
     const result = await SellerCollection.deleteOne(query);
 
     sendResponse(res, {
@@ -58,6 +70,16 @@ const updateSeller = catchAsync(async (req, res) => {
     const sellerData = req.body;
     const id = req.params.id;
     const filter = { _id: new ObjectId(id) };
+
+    const sellerRequest = await SellerCollection.findOne(filter);
+    if (!sellerRequest) {
+        throw new AppError(404, "Seller request not found");
+    }
+
+    if (req.user.role !== "admin" && sellerRequest.sellerEmail !== req.user.email) {
+        throw new AppError(403, "You do not have permission to update this request");
+    }
+
     const updatedDoc = {
         $set: {
             sellerName: sellerData.sellerName,
@@ -65,7 +87,9 @@ const updateSeller = catchAsync(async (req, res) => {
             other: sellerData.other,
             address: sellerData.address,
             reason: sellerData.reason,
-            imageUrl: sellerData.imageUrl
+            imageUrl: sellerData.imageUrl,
+            adminIsApproved: "Pending",
+            decline: ""
         }
     }
 
@@ -104,17 +128,72 @@ const sellerDecline = catchAsync(async (req, res) => {
     const id = req.params.id;
     const filter = { _id: new ObjectId(id) }
 
+    const sellerRequest = await SellerCollection.findOne(filter);
+    if (!sellerRequest) {
+        return res.status(404).send({ success: false, message: "Seller request not found" });
+    }
+
     const updatedDoc = {
         $set: {
+            adminIsApproved: "Declined",
             decline: declineMessage.inputValue
         }
     }
     const result = await SellerCollection.updateOne(filter, updatedDoc)
 
+    // Send email
+    await sendSellerStatusEmail(
+        sellerRequest.sellerEmail,
+        sellerRequest.sellerName,
+        "Declined",
+        declineMessage.inputValue
+    );
+
     sendResponse(res, {
         statusCode: 201,
         success: true,
         message: "seller decline message success",
+        data: result
+    });
+});
+
+const approveSeller = catchAsync(async (req, res) => {
+    const id = req.params.id;
+    const filter = { _id: new ObjectId(id) };
+
+    const sellerRequest = await SellerCollection.findOne(filter);
+    if (!sellerRequest) {
+        return res.status(404).send({ success: false, message: "Seller request not found" });
+    }
+
+
+    const updatedDoc = {
+        $set: {
+            adminIsApproved: "Approved",
+            decline: ""
+        }
+    };
+    const result = await SellerCollection.updateOne(filter, updatedDoc);
+
+
+    const userFilter = { email: sellerRequest.sellerEmail };
+    const userUpdate = {
+        $set: {
+            role: "host"
+        }
+    };
+    await UserCollection.updateOne(userFilter, userUpdate);
+
+    await sendSellerStatusEmail(
+        sellerRequest.sellerEmail,
+        sellerRequest.sellerName,
+        "Approved"
+    );
+
+    sendResponse(res, {
+        statusCode: 200,
+        success: true,
+        message: "seller approved success",
         data: result
     });
 });
@@ -126,7 +205,8 @@ const SellerController = {
     deleteSeller,
     updateSeller,
     createSeller,
-    sellerDecline
+    sellerDecline,
+    approveSeller
 };
 
 module.exports = SellerController;
